+++ spec/features/restore_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe "imap-backup restore", :docker, type: :aruba do
   include_context "message-fixtures"
 
   let(:account_config) { test_server_connection_parameters.merge(folders: [folder]) }
-  let(:folder) { "my-stuff" }
+  let(:folder) { "stuff-to-restore" }
   let(:messages_as_mbox) do
     to_mbox_entry(**message_one) + to_mbox_entry(**message_two)
   end
@@ -12,10 +12,10 @@ RSpec.describe "imap-backup restore", :docker, type: :aruba do
     [message_as_server_message(**message_one), message_as_server_message(**message_two)]
   end
   let(:uid_validity) { 1234 }
-  let(:email) { test_server_connection_parameters[:username] }
   let(:config_options) { {accounts: [account_config]} }
+  let(:email) { account_config[:username] }
 
-  let!(:pre) {}
+  let!(:pre) { test_server.warn_about_non_default_folders }
   let!(:setup) do
     create_config(**config_options)
     create_local_folder email: email, folder: folder, uid_validity: uid_validity
@@ -23,9 +23,8 @@ RSpec.describe "imap-backup restore", :docker, type: :aruba do
     append_local(
       email: email, folder: folder, flags: [:Draft, :$NON_SYSTEM_FLAG], **message_two
     )
-
-    run_command_and_stop("imap-backup restore #{email}")
   end
+  let(:run_command) { run_command_and_stop("imap-backup restore #{email}") }
   let(:cleanup) do
     test_server.delete_folder folder
     test_server.disconnect
@@ -35,11 +34,15 @@ RSpec.describe "imap-backup restore", :docker, type: :aruba do
 
   context "when the folder doesn't exist" do
     it "restores messages" do
+      run_command
+
       messages = test_server.folder_messages(folder).map { |m| server_message_to_body(m) }
       expect(messages).to eq(messages_as_server_messages)
     end
 
     it "restores flags" do
+      run_command
+
       messages = test_server.folder_messages(folder)
       flags = messages.map { |m| m["FLAGS"] }
 
@@ -47,12 +50,16 @@ RSpec.describe "imap-backup restore", :docker, type: :aruba do
     end
 
     it "updates local uids to match the new server ones" do
+      run_command
+
       updated_imap_content = imap_parsed(email, folder)
       stored_uids = updated_imap_content[:messages].map { |m| m[:uid] }
       expect(test_server.folder_uids(folder)).to eq(stored_uids)
     end
 
     it "sets the backup uid_validity to match the new folder" do
+      run_command
+
       updated_imap_content = imap_parsed(email, folder)
       expect(updated_imap_content[:uid_validity]).
         to eq(test_server.folder_uid_validity(folder))
@@ -63,10 +70,11 @@ RSpec.describe "imap-backup restore", :docker, type: :aruba do
     let(:email3) { test_server.send_email folder, **message_three }
 
     context "when the uid_validity matches" do
-      let(:pre) do
+      let(:setup) do
         test_server.create_folder folder
         email3
         uid_validity
+        super()
       end
       let(:messages_as_server_messages) do
         [
@@ -78,6 +86,8 @@ RSpec.describe "imap-backup restore", :docker, type: :aruba do
       let(:uid_validity) { test_server.folder_uid_validity(folder) }
 
       it "appends to the existing folder" do
+        run_command
+
         messages = test_server.folder_messages(folder).map { |m| server_message_to_body(m) }
         expect(messages).to eq(messages_as_server_messages)
       end
@@ -85,17 +95,22 @@ RSpec.describe "imap-backup restore", :docker, type: :aruba do
 
     context "when the uid_validity doesn't match" do
       context "when the folder is empty" do
-        let(:pre) do
+        let(:setup) do
           test_server.create_folder folder
+          super()
         end
 
         it "sets the backup uid_validity to match the folder" do
+          run_command
+
           updated_imap_content = imap_parsed(email, folder)
           expect(updated_imap_content[:uid_validity]).
             to eq(test_server.folder_uid_validity(folder))
         end
 
         it "uploads to the new folder" do
+          run_command
+
           messages = test_server.folder_messages(folder).map do |m|
             server_message_to_body(m)
           end
@@ -105,9 +120,10 @@ RSpec.describe "imap-backup restore", :docker, type: :aruba do
 
       context "when the folder has content" do
         let(:new_folder) { "#{folder}-#{uid_validity}" }
-        let(:pre) do
+        let(:setup) do
           test_server.create_folder folder
           email3
+          super()
         end
         let(:cleanup) do
           test_server.delete_folder new_folder
@@ -115,10 +131,14 @@ RSpec.describe "imap-backup restore", :docker, type: :aruba do
         end
 
         it "renames the backup" do
+          run_command
+
           expect(mbox_content(email, new_folder)).to eq(messages_as_mbox)
         end
 
         it "leaves the existing folder as is" do
+          run_command
+
           messages = test_server.folder_messages(folder).map do |m|
             server_message_to_body(m)
           end
@@ -126,16 +146,22 @@ RSpec.describe "imap-backup restore", :docker, type: :aruba do
         end
 
         it "creates the new folder" do
+          run_command
+
           expect(test_server.folders.map(&:name)).to include(new_folder)
         end
 
         it "sets the backup uid_validity to match the new folder" do
+          run_command
+
           updated_imap_content = imap_parsed(email, new_folder)
           expect(updated_imap_content[:uid_validity]).
             to eq(test_server.folder_uid_validity(new_folder))
         end
 
         it "uploads to the new folder" do
+          run_command
+
           messages = test_server.folder_messages(new_folder).map do |m|
             server_message_to_body(m)
           end
@@ -154,11 +180,11 @@ RSpec.describe "imap-backup restore", :docker, type: :aruba do
       create_config accounts: [account_config]
       create_local_folder email: email, folder: folder, uid_validity: uid_validity
       append_local email: email, folder: folder, **msg_iso8859
-
-      run_command_and_stop("imap-backup restore #{email}")
     end
 
     it "maintains encodings" do
+      run_command
+
       message =
         test_server.folder_messages(folder).
         first["BODY[]"]
@@ -189,13 +215,23 @@ RSpec.describe "imap-backup restore", :docker, type: :aruba do
         **message_one
       )
     end
-
-    it "does not raise any errors" do
+    let(:run_command) do
       run_command_and_stop(
         "imap-backup restore #{email} --config #{custom_config_path}"
       )
+    end
+
+    it "does not raise any errors" do
+      run_command
 
       expect(last_command_started).to have_exit_status(0)
+    end
+
+    it "restores messages" do
+      run_command
+
+      messages = test_server.folder_messages(folder).map { |m| server_message_to_body(m) }
+      expect(messages).to eq([message_as_server_message(**message_one)])
     end
   end
 end