+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## next - next
+
+### Added
+
+* A new `single backup` command that backs up an account
+  without the need for a configuration file.
+
 ## 13.4.0 - 2023-10-09
 
 ### Changed